+++ packages/react/src/index.ts
@@ -4,7 +4,6 @@ import curious from './curious';
 import Link from './Link';
 import Navigator from './Navigator';
 import Provider from './Provider';
-import Redirect from './Redirect';
 
 export {
   Active,
@@ -12,6 +11,5 @@ export {
   curious,
   Link,
   Navigator,
-  Provider,
-  Redirect
+  Provider
 };