+++ packages/react/types/index.d.ts
@@ -4,5 +4,4 @@ import curious from './curious';
 import Link from './Link';
 import Navigator from './Navigator';
 import Provider from './Provider';
-import Redirect from './Redirect';
-export { Active, Block, curious, Link, Navigator, Provider, Redirect };
+export { Active, Block, curious, Link, Navigator, Provider };